Adicionar empty-image na collection dos roteiros
