+++ macroChallengeApp/macroChallengeApp/Flows/Profile/UIViewCell/ProfileCollectionViewCell.swift
@@ -99,19 +99,14 @@ extension ProfileCollectionViewCell {
         }
     }
     
-    func setupImage(category: String) {
-        let path = category
-        let imageNew = UIImage(contentsOfFile: SaveImagecontroller.getFilePath(fileName: path))
-        self.roadmapImage.image = imageNew
-//        if category == "Beach" {
-//            self.roadmapImage.image = designSystem.imagesDefault.beach[Int.random(in: 0..<designSystem.imagesDefault.beach.count)]
-//        } else if category == "Mountain" {
-//            self.roadmapImage.image = designSystem.imagesDefault.mountain[Int.random(in: 0..<designSystem.imagesDefault.mountain.count)]
-//        } else if category == "City" {
-//            self.roadmapImage.image = designSystem.imagesDefault.city[Int.random(in: 0..<designSystem.imagesDefault.city.count)]
-//        } else {
-//            self.roadmapImage.image = designSystem.imagesDefault.camp[Int.random(in: 0..<designSystem.imagesDefault.camp.count)]
-//        }
+    func setupImage(imageId: String, category: String) {
+        if imageId == "defaultCover" {
+            self.roadmapImage.image = UIImage(named: "\(imageId)\(category)")
+        } else {
+            let path = imageId
+            let imageNew = UIImage(contentsOfFile: SaveImagecontroller.getFilePath(fileName: path))
+            self.roadmapImage.image = imageNew
+        }
     }
     
     func setupConstraints() {
